all tests passed

+++ lib/src/crypto/dartRandom.dart
@@ -23,7 +23,7 @@ class DartRandom implements SecureRandom {
     /// forcing remainingBits to be calculate with bitLength
     int remainingBits = (bitLength - main.bitLength);
     int additional = remainingBits < 4
-        ? dartRandom.nextInt(pow(2, remainingBits))
+        ? dartRandom.nextInt(pow(2, remainingBits) as int)
         : remainingBits;
     BigInt additionalBit = (new BigInt.from(additional) << (fullBytes * 8));
     BigInt result = main + additionalBit;
@@ -41,13 +41,13 @@ class DartRandom implements SecureRandom {
   }
 
   @override
-  int nextUint16() => dartRandom.nextInt(pow(2, 32));
+  int nextUint16() => dartRandom.nextInt(pow(2, 32) as int);
 
   @override
-  int nextUint32() => dartRandom.nextInt(pow(2, 32));
+  int nextUint32() => dartRandom.nextInt(pow(2, 32) as int);
 
   @override
-  int nextUint8() => dartRandom.nextInt(pow(2, 8));
+  int nextUint8() => dartRandom.nextInt(pow(2, 8) as int);
 
   @override
   void seed(CipherParameters params) {
